+++ community/boilerplates/remote-schemas/aws-lambda/nodejs/.gitignore
@@ -1,2 +1,3 @@
 node_modules/
 *.zip
+package-lock.json
