+++ community/boilerplates/remote-schemas/azure-functions/nodejs/.gitignore
@@ -1,2 +1,3 @@
 node_modules/
 *.zip
+package-lock.json
