+++ community/boilerplates/remote-schemas/google-cloud-functions/nodejs/.gitignore
@@ -1,2 +1,3 @@
 node_modules/
 *.zip
+package-lock.json
