--- community/boilerplates/remote-schemas/zeit-now/nodejs/express-graphqljs-sequelize/Dockerfile
@@ -1,11 +1,0 @@
-FROM node:8
-
-WORKDIR /server
-
-COPY ./package.json /server/
-
-RUN npm install
-
-COPY . /server/
-
-CMD ["npm", "start"]
