--- community/boilerplates/remote-schemas/zeit-now/nodejs/express-graphqljs-sequelize/now.json
@@ -1,3 +1,0 @@
-{
- "version": 1
-}
