--- community/boilerplates/remote-schemas/zeit-now/nodejs/express-graphqljs-sequelize/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "nodejs-express-gql-server",
-  "version": "1.0.0",
-  "description": "A GraphQL server boilerplate for NodeJS-Express using graphql-js library",
-  "main": "server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node -r esm server.js"
-  },
-  "author": "Hasura",
-  "license": "MIT",
-  "dependencies": {
-    "express": "^4.16.4",
-    "express-graphql": "^0.7.1",
-    "graphql": "^14.0.2",
-    "pg": "^7.7.1",
-    "sequelize": "^4.42.0"
-  },
-  "devDependencies": {
-    "esm": "^3.0.84"
-  }
-}
