+++ community/boilerplates/remote-schemas/zeit-now/nodejs/now.json
