+++ community/boilerplates/remote-schemas/zeit-now/nodejs/package.json
@@ -4,18 +4,13 @@
   "description": "A GraphQL server boilerplate written in NodeJS using and Apollo Server",
   "main": "server.js",
   "scripts": {
-    "start": "node -r esm server.js"
+    "start": "node server.js"
   },
   "author": "Hasura",
   "license": "MIT",
   "dependencies": {
     "apollo-server": "^2.2.1",
     "graphql": "^14.0.2",
-    "graphql-tools": "^4.0.3",
-    "pg": "^7.7.1",
-    "sequelize": "^4.42.0"
-  },
-  "devDependencies": {
-    "esm": "^3.0.84"
+    "graphql-tools": "^4.0.3"
   }
 }
